Save image names to file
