+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
 client/public/bundle.js
-seed/img
+seed/retrieveImgs/img
