+++ seed/retrieveImgs/config.js
